Cleaned up and removed obsolete code
Removed oauth app. For new projects that wish to use oauth, look into the
django-allauth third party package.

Also removed a couple things that we don't really use, but are easy to add back
in if we want: package.json and setup.cfg.

Removed .travis.yml as we no longer use github.