removed isort and warnings from tox config
isort was incorrectly detecting django as a first party package and
insisted several correct imports were incorrect. I don't have time to
figure it out right now or fix the bug, so I'm removing isort for the
moment.

The problem with running with warnings as errors is that our tests can
fail due to deprication warnings in third party libraries. Our tests
passing shouldn't depend on the internals of third party libraries, so
I'm taking this out too.

Both are still something that would be nice to run once in a while ,but
not on every tox invocation.